requirement.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
carehare==1.0.1
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-rabbitmq==4.0.0
channels-redis==3.3.1
charset-normalizer==2.0.9
constantly==15.1.0
cryptography==36.0.0
daphne==3.0.2
Deprecated==1.2.13
Django==3.2.9
django-cleanup==5.2.0
django-cors-headers==3.10.0
django-redis==5.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.3
oauthlib==3.1.1
pamqp==3.0.1
Pillow==8.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
pytz==2021.3
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing_extensions==4.0.0
urllib3==1.26.7
wrapt==1.13.3
zope.interface==5.4.0